fix: Don't set prefix/version to None in test message conversion
The issue was that convert_to_test_message() was setting prefix and version
to None when they weren't present in the input signal_body. The encoder
expects these fields to either have valid values or not be present at all.

Solution: Only add prefix/version to the test_message dict if they exist
in the original signal_body. The encoder will add defaults (WF/1) if needed.